% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make groups}
\usage{
make_groups(names)
}
\arguments{
\item{names}{A list of names}
}
\value{
A two column matrix of names
}
\description{
Make groups of two randomly from a list of names
}
