% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRawdata.R
\name{read.irradiance}
\alias{read.irradiance}
\title{Read sediment irradiance raw data}
\usage{
read.irradiance(File = "")
}
\arguments{
\item{File}{The folder that you save your txt files}
}
\value{
A dataframe named merge.data
}
\description{
Read all the txt files in the folder and merge them into a data frame
}
